Modified Java file

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
-    public int calculate(int mass, double height) {
+    public double calculate(int mass, double height) {
         double index = mass / (height * height);
-        return (int) index;
+        return index;
     }
 }
